requirements.txt
fastapi~=0.115.6
python-dotenv~=1.0.1
pydantic-settings~=2.7.0
pydantic~=2.10.4
moviepy~=2.1.1
aiofiles~=24.1.0
google-api-python-client~=2.156.0
tenacity~=9.0.0
openai~=1.58.1
uvicorn
deepgram-sdk
python-multipart
httpx~=0.28.1